Misc improvements/fixes (issues discovered when testing SNT  w/ Fiji's new launcher)
- Fix NPE
- Fix columns in measurements table without units
- Make RootAngleAnalyzer use all available LUTs
- Misc

+++ src/main/java/sc/fiji/snt/plugin/ShollAnalysisImgCommonCmd.java
@@ -1119,7 +1119,7 @@ public class ShollAnalysisImgCommonCmd extends DynamicCommand {
 			final LinearProfileStats lStats = new LinearProfileStats(profile);
 			lStats.setLogger(logger);
 			if (primaryBranches > 0 ) {
-				lStats.setPrimaryBranches((int)primaryBranches);
+				lStats.setPrimaryBranches(primaryBranches);
 			}
 
 			if (polynomialChoice.contains("Best")) {
